Prevent EMLProcessor from re-inserting identifier blank nodes
We use blank nodes for schema:identifier statements. If we reinsert the same dataset, we would get what are effectively duplicate blank nodes for things like identifier. ie If we inserted _twice_ a dataset where PersonA is a creator with ORCID O, we'd have to blank nodes for for PersonA, both with the value O. This isn't wrong but it'll make queries harder. So now we query first and don't re-insert the schema:identifier blank node (and its triples) if an equivalent set of triples already exist.